inverse-destination-popularity/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0
fastparquet>=2023.1.0
geopandas>=0.12.0
shapely>=2.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
pyspark>=3.3.0
requests>=2.28.0
google-cloud-storage>=2.10.0

